pyproject.toml
[tool.poetry]
name = "bo_nalog"
version = "0.1.0"
description = ""
authors = ["Max Zakirov <zakirov_mv@srvhub.ru>"]

[[tool.poetry.source]]
name = "default"
url = "https://pypi.org/simple"
default = true

[[tool.poetry.source]]
name = "srvhub"
url = "http://pypi.srvhub.tools/simple"
secondary = true

[tool.poetry.dependencies]
python = "3.10.0"
django = "3.2.8"
psycopg2 = "2.9.1"
celery = "5.1.2"
requests = "2.26.0"
djangorestframework = "3.12.4"
sentry-sdk = "1.4.3"
jsondiff = "1.3.0"
xmltodict = "0.12.0"
python-redis-lock = "3.7.0"
django-celery-beat = "2.2.1"
django-redis = "5.0.0"
django-prometheus = "2.1.0"
requests-toolbelt = "0.9.1"
django-json-widget = "1.1.1"
django-dynamic-raw-id = "2.8"
djnewsletter = {version = "^2.3.0", source = "srvhub"}
dj-model-utils = {version = "^0.1.0", source = "srvhub"}
django-environ = "0.7.0"
appversion = {version = "^0.1.4", source = "srvhub"}
envparse = "0.2.0"
uvicorn = "0.15.0"
random-user-agent = "1.0.1"
aiohttp = "3.8.0"

[tool.poetry.dev-dependencies]
mock = "4.0.3"
coverage = "6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"